requirements.txt
langchain_google_genai
langchain_core
streamlit
python-dotenv
langchain_community
langserve
fastapi
uvicorn
sse_starlette